+++ R/gmte_aalen.R
@@ -230,8 +230,7 @@ gmte_aalen = function(Y_t0,Y_t1,Y_d,T,G,Z,D,Nsim=100,alpha=0.05,verbose=FALSE)
 	FullCombined[10,] = RGMTE_MR_CAT
 
 	colnames(FullCombined) = c("Est","SE","EstP","Qstat","Qp","Combine?")
-	rownames(FullCombined) = c("CAT","GMTE0","GMTE1","RGMTE","MR",
-	                           "RGMTE_MR","RGMTE_CAT","MR_CAT","GMTE1_CAT","RGMTE_MR_CAT")
+	FullCombined = cbind(Model=c("CAT","GMTE0","GMTE1","RGMTE","MR","RGMTE_MR","RGMTE_CAT","MR_CAT","GMTE1_CAT","RGMTE_MR_CAT"), as.data.frame(FullCombined))
 
 	cat("\nResults:\n")
 	print(FullCombined)